Fix bugs
This includes also changing the way tge <modsCollection> element is generated, in order
to be able to add the namespace attributes and schema without xslt doing weird stuff
(adding the name space in the <modsCollection> affects the namespaces used
by xslt while processing - or something along those lines. So had to output the
<modsCollection> element and attributes via <xsl:text> instead of just including in
the xslt xml.