Implement <typeOfResource>
NOTE: we use the itemType given in the metadata to populate
<typeOfResource>. However, the CUL specs require that, if there
is not itemType, that a value be generated based on the type of the
associated file(s). THIS HAS NOT BEEN IMPLEMENTED. It is probably
easier to create itemType in the remediated metadata, based on the
type of the associated files.