Change filename extension from .jpg to .tif
This is done because the files loaded into Omeka are jpg files,
but the original files from the imaging lab are tif files.
Note that the extension for files not generated by the imaging
lab were not changed since these file will not be stored in
Fedora, and therefore the associated metadata will not be used.