Remove prefix "itemContainer_-_" from elements containing it
In the first four set of remediated data (Frances Perkins, Melting Pot,
Ballets Russes, George Plimpton), the elements in the remediated metadata
started with the prefix "item_-_" (for example, <item_-_MODS_-_DigitalOrigin>),
but in this set of remediated metadata, the prefix is "itemContainer_-_item_-_"
(for example, <itemContainer_-_item_-_MODS_-_DigitalOrigin>). So the prefix
has to be normalized to what was used in the first four sets of metadata.