Add new dates that were missing in remediated metadata
When the George Plimpton MODS was first generated from the prepped
remediated metadata, an empty <originInfo/> was generated. Turns out
3 items were missing a start date (<originInfo> can contain other
info besides dates, which is why it was generatd for 2 out of the 3
items without dates).
Metadata group fixed the remediated metadata, and the changes were
inserted via this commit.