Remove empty elements from MODS file
With the XLST stylesheet as written, if a wrapper elements turns out
to have no subelements inside of it, xsltproc will still create an
empty element (such as <subject/>). One good side effect is that it
allows one to see if some info is missing. However, it also means that
these empty element have to be removed by hand.