Remove "b"s appended to some of the filenames, Melting Pot
Melting Pot:
For some reason, some of the filenames for the images
loaded into Omeka have a "b" appended to the filename - for example
2_04_08b.jpg. In each of these cases, the image contained in the JPG
file is the same as the image in the associated
(without a "b" appended to the filename) TIFF file. So these appended
"b"s were removed.