Add generation of project URL, if needed
The Project URL is the resolver URL as found in CLIO. This information
is not stored in Omeka, and needs to be generated. In the first few
colelction that were remediated, this project url was included in the
remediated metadata. However, it was not included in the Music at
Columbia colelction. Therefore, the code was update to handle
generation of the project url if needed.