Updated metadata for Frances Perkins
Changes include a few column name changes, the addtion of whole
and part for a title, title attribute for secondary title.
Also, the remediated metadata will no longer be prepped, so this
version still has the .jpg extension, which the XSLT stylesheet
will have to handle, as well as non-normalized filenames, which will
also have to be done by the XSLT stylesheet.