Differentiate between main title and secondary title(s)
There can only be one <titleInfo> top-level element without an
attribute -- this is a CUL-software restriction. This will be
considered the main title. The other titles are required to have
the type attribute set in its top-level <titleInfo> -- again, this
is a CUL-software restriction, which is used by said software to
differentiates the main title from the "secondary" titles.