Add regular expression processing for <item_-_MODS_-_FormGenre..>
Now that there is an associated authority element of each one
of the above, need to use regular expressions to parse out
the correct element. This commit contains a matches() call
that seems to work. Need to test it out more, and then move it
to the other affected elements.